Remove yarn.lock file
The yarn.lock file was deleted to help manage dependencies more efficiently. This file was auto-generated and contained various versions of each dependency, which might cause conflicting versions to be installed. With the removal of this file, the project will always install the latest versions specified in the package.json file.